third_party/WebKit/LayoutTests/fast/dom/idl-union-type-unittest-expected.txt

IDL union types unittests

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Tests for attributes
PASS unionTypesTest.doubleOrStringOrStringArrayAttribute is 0
PASS unionTypesTest.doubleOrStringOrStringArrayAttribute is 3.14
PASS unionTypesTest.doubleOrStringOrStringArrayAttribute is "foo"
PASS unionTypesTest.doubleOrStringOrStringArrayAttribute is "undefined"
PASS unionTypesTest.doubleOrStringOrStringArrayAttribute is "null"
PASS unionTypesTest.doubleOrStringOrStringArrayAttribute is "[object Object]"
PASS JSON.stringify(unionTypesTest.doubleOrStringOrStringArrayAttribute) is "[]"
PASS JSON.stringify(unionTypesTest.doubleOrStringOrStringArrayAttribute) is "[\"a\",\"3.14\"]"

Tests for method arguments
PASS unionTypesTest.doubleOrStringArg(3.14) is "double is passed: 3.14"
PASS unionTypesTest.doubleOrStringArg("foo") is "string is passed: foo"
PASS unionTypesTest.doubleOrStringArg(undefined) is "string is passed: undefined"
PASS unionTypesTest.doubleOrStringArg(null) is "string is passed: null"
PASS unionTypesTest.doubleOrStringArg({}) is "string is passed: [object Object]"
PASS unionTypesTest.doubleOrStringArg([]) is "string is passed: "
PASS typeof unionTypesTest.doubleOrStringArg(new Date) is "string"
PASS unionTypesTest.doubleOrStringArg() threw exception TypeError: Failed to execute 'doubleOrStringArg' on 'UnionTypesTest': 1 argument required, but only 0 present..

PASS unionTypesTest.doubleOrInternalEnumArg(3.14) is "double is passed: 3.14"
PASS unionTypesTest.doubleOrInternalEnumArg("foo") is "InternalEnum is passed: foo"
PASS unionTypesTest.doubleOrInternalEnumArg("invalid") threw exception TypeError: Failed to execute 'doubleOrInternalEnumArg' on 'UnionTypesTest': The provided value 'invalid' is not a valid enum value of type InternalEnum..
PASS unionTypesTest.doubleOrInternalEnumArg(null) threw exception TypeError: Failed to execute 'doubleOrInternalEnumArg' on 'UnionTypesTest': The provided value 'null' is not a valid enum value of type InternalEnum..
PASS unionTypesTest.doubleOrInternalEnumArg({}) threw exception TypeError: Failed to execute 'doubleOrInternalEnumArg' on 'UnionTypesTest': The provided value '[object Object]' is not a valid enum value of type InternalEnum..
PASS unionTypesTest.doubleOrInternalEnumArg([]) threw exception TypeError: Failed to execute 'doubleOrInternalEnumArg' on 'UnionTypesTest': The provided value '' is not a valid enum value of type InternalEnum..

PASS unionTypesTest.doubleOrStringArrayArg([]) is ""
PASS unionTypesTest.doubleOrStringArrayArg([3.14, "foo"]) is "double: 3.14, string: foo"
PASS unionTypesTest.doubleOrStringArrayArg([1, "foo", "bar", 2]) is "double: 1, string: foo, string: bar, double: 2"
PASS unionTypesTest.doubleOrStringArrayArg([null, undefined, {}, []]) is "string: null, string: undefined, string: [object Object], string: "
PASS unionTypesTest.doubleOrStringArrayArg(null) threw exception TypeError: Failed to execute 'doubleOrStringArrayArg' on 'UnionTypesTest': The 1st argument is neither an array, nor does it have indexed properties..
PASS unionTypesTest.doubleOrStringArrayArg(undefined) threw exception TypeError: Failed to execute 'doubleOrStringArrayArg' on 'UnionTypesTest': The 1st argument is neither an array, nor does it have indexed properties..
PASS unionTypesTest.doubleOrStringArrayArg({}) threw exception TypeError: Failed to execute 'doubleOrStringArrayArg' on 'UnionTypesTest': The 1st argument is neither an array, nor does it have indexed properties..

PASS unionTypesTest.doubleOrStringSequenceArg([]) is ""
PASS unionTypesTest.doubleOrStringSequenceArg([3.14, "foo"]) is "double: 3.14, string: foo"
PASS unionTypesTest.doubleOrStringSequenceArg([1, "foo", "bar", 2]) is "double: 1, string: foo, string: bar, double: 2"
PASS unionTypesTest.doubleOrStringSequenceArg([null, undefined, {}, []]) is "string: null, string: undefined, string: [object Object], string: "
PASS unionTypesTest.doubleOrStringSequenceArg(null) threw exception TypeError: Failed to execute 'doubleOrStringSequenceArg' on 'UnionTypesTest': The 1st argument is neither an array, nor does it have indexed properties..
PASS unionTypesTest.doubleOrStringSequenceArg(undefined) threw exception TypeError: Failed to execute 'doubleOrStringSequenceArg' on 'UnionTypesTest': The 1st argument is neither an array, nor does it have indexed properties..
PASS unionTypesTest.doubleOrStringSequenceArg({}) threw exception TypeError: Failed to execute 'doubleOrStringSequenceArg' on 'UnionTypesTest': The 1st argument is neither an array, nor does it have indexed properties..

Tests for method arguments with defaults
PASS unionTypesTest.doubleOrStringDefaultDoubleArg() is "double is passed: 3.14"
PASS unionTypesTest.doubleOrStringDefaultDoubleArg(undefined) is "double is passed: 3.14"
PASS unionTypesTest.doubleOrStringDefaultDoubleArg("foo") is "string is passed: foo"
PASS unionTypesTest.doubleOrStringDefaultStringArg() is "string is passed: foo"
PASS unionTypesTest.doubleOrStringDefaultStringArg(undefined) is "string is passed: foo"
PASS unionTypesTest.doubleOrStringDefaultStringArg(3.14) is "double is passed: 3.14"
PASS unionTypesTest.doubleOrStringDefaultNullArg() is "null is passed"
PASS unionTypesTest.doubleOrStringDefaultNullArg(undefined) is "null is passed"
PASS unionTypesTest.doubleOrStringDefaultNullArg(null) is "null is passed"
PASS unionTypesTest.doubleOrStringDefaultNullArg(3.14) is "double is passed: 3.14"
PASS unionTypesTest.doubleOrStringDefaultNullArg("foo") is "string is passed: foo"

Tests for nullable method arguments
PASS unionTypesTest.nodeListOrElementOrNullArg(undefined) is "null or undefined is passed"
PASS unionTypesTest.nodeListOrElementOrNullArg(null) is "null or undefined is passed"
PASS unionTypesTest.nodeListOrElementOrNullArg(nodeList) is "nodelist is passed"
PASS unionTypesTest.nodeListOrElementOrNullArg(element) is "element is passed"
PASS unionTypesTest.nodeListOrElementOrNullArg(3.14) threw exception TypeError: Failed to execute 'nodeListOrElementOrNullArg' on 'UnionTypesTest': The provided value is not of type '(NodeList or Element)'.
PASS unionTypesTest.nodeListOrElementOrNullArg("foo") threw exception TypeError: Failed to execute 'nodeListOrElementOrNullArg' on 'UnionTypesTest': The provided value is not of type '(NodeList or Element)'.
PASS unionTypesTest.nodeListOrElementOrNullArg({}) threw exception TypeError: Failed to execute 'nodeListOrElementOrNullArg' on 'UnionTypesTest': The provided value is not of type '(NodeList or Element)'.
PASS unionTypesTest.nodeListOrElementOrNullArg([]) threw exception TypeError: Failed to execute 'nodeListOrElementOrNullArg' on 'UnionTypesTest': The provided value is not of type '(NodeList or Element)'.
PASS unionTypesTest.nodeListOrElementArg(undefined) threw exception TypeError: Failed to execute 'nodeListOrElementArg' on 'UnionTypesTest': The provided value is not of type '(NodeList or Element)'.
PASS unionTypesTest.nodeListOrElementArg(null) threw exception TypeError: Failed to execute 'nodeListOrElementArg' on 'UnionTypesTest': The provided value is not of type '(NodeList or Element)'.
PASS unionTypesTest.nodeListOrElementArg(nodeList) is "nodelist is passed"
PASS unionTypesTest.nodeListOrElementArg(element) is "element is passed"

Tests for method arguments with array/sequence members
PASS unionTypesTest.doubleOrStringOrStringArrayArg() threw exception TypeError: Failed to execute 'doubleOrStringOrStringArrayArg' on 'UnionTypesTest': 1 argument required, but only 0 present..
PASS unionTypesTest.doubleOrStringOrStringArrayArg("abc") is "string: abc"
PASS unionTypesTest.doubleOrStringOrStringArrayArg(3.14) is "double: 3.14"
PASS unionTypesTest.doubleOrStringOrStringArrayArg(true) is "string: true"
PASS unionTypesTest.doubleOrStringOrStringArrayArg(null) is "string: null"
PASS unionTypesTest.doubleOrStringOrStringArrayArg(undefined) is "string: undefined"
PASS unionTypesTest.doubleOrStringOrStringArrayArg([]) is "array: []"
PASS unionTypesTest.doubleOrStringOrStringArrayArg(["a", 3.14]) is "array: [a, 3.14]"
PASS unionTypesTest.doubleOrStringOrStringArrayArg(new Array(3)) is "array: [undefined, undefined, undefined]"
PASS unionTypesTest.doubleOrStringOrStringSequenceArg() threw exception TypeError: Failed to execute 'doubleOrStringOrStringSequenceArg' on 'UnionTypesTest': 1 argument required, but only 0 present..
PASS unionTypesTest.doubleOrStringOrStringSequenceArg("abc") is "string: abc"
PASS unionTypesTest.doubleOrStringOrStringSequenceArg(3.14) is "double: 3.14"
PASS unionTypesTest.doubleOrStringOrStringSequenceArg(true) is "string: true"
PASS unionTypesTest.doubleOrStringOrStringSequenceArg(null) is "string: null"
PASS unionTypesTest.doubleOrStringOrStringSequenceArg(undefined) is "string: undefined"
PASS unionTypesTest.doubleOrStringOrStringSequenceArg([]) is "sequence: []"
PASS unionTypesTest.doubleOrStringOrStringSequenceArg(["a", 3.14]) is "sequence: [a, 3.14]"
PASS unionTypesTest.doubleOrStringOrStringSequenceArg(new Array(3)) is "sequence: [undefined, undefined, undefined]"

Tests for nullable method arguments with array/sequence members
PASS unionTypesTest.doubleOrStringOrStringArrayNullableArg() threw exception TypeError: Failed to execute 'doubleOrStringOrStringArrayNullableArg' on 'UnionTypesTest': 1 argument required, but only 0 present..
PASS unionTypesTest.doubleOrStringOrStringArrayNullableArg("abc") is "string: abc"
PASS unionTypesTest.doubleOrStringOrStringArrayNullableArg(3.14) is "double: 3.14"
PASS unionTypesTest.doubleOrStringOrStringArrayNullableArg(true) is "string: true"
PASS unionTypesTest.doubleOrStringOrStringArrayNullableArg(null) is "null"
PASS unionTypesTest.doubleOrStringOrStringArrayNullableArg(undefined) is "null"
PASS unionTypesTest.doubleOrStringOrStringArrayNullableArg([]) is "array: []"
PASS unionTypesTest.doubleOrStringOrStringArrayNullableArg(["a", 3.14]) is "array: [a, 3.14]"
PASS unionTypesTest.doubleOrStringOrStringArrayNullableArg(new Array(3)) is "array: [undefined, undefined, undefined]"
PASS unionTypesTest.doubleOrStringOrStringSequenceNullableArg() threw exception TypeError: Failed to execute 'doubleOrStringOrStringSequenceNullableArg' on 'UnionTypesTest': 1 argument required, but only 0 present..
PASS unionTypesTest.doubleOrStringOrStringSequenceNullableArg("abc") is "string: abc"
PASS unionTypesTest.doubleOrStringOrStringSequenceNullableArg(3.14) is "double: 3.14"
PASS unionTypesTest.doubleOrStringOrStringSequenceNullableArg(true) is "string: true"
PASS unionTypesTest.doubleOrStringOrStringSequenceNullableArg(null) is "null"
PASS unionTypesTest.doubleOrStringOrStringSequenceNullableArg(undefined) is "null"
PASS unionTypesTest.doubleOrStringOrStringSequenceNullableArg([]) is "sequence: []"
PASS unionTypesTest.doubleOrStringOrStringSequenceNullableArg(["a", 3.14]) is "sequence: [a, 3.14]"
PASS unionTypesTest.doubleOrStringOrStringSequenceNullableArg(new Array(3)) is "sequence: [undefined, undefined, undefined]"
PASS successfullyParsed is true

TEST COMPLETE
